refactor: 페이지별 로직 분리로 App.tsx 복잡도 개선
